Finished paging, printing hexadecimal correctly.

+++ OS-Main/libc/include/stdio.h
@@ -19,7 +19,7 @@ void sendToGets(char);
 int printf(const char* __restrict, ...);
 int putchar(int);
 int puts(const char*);
-char* itoa(int value, char * str, int base);
+char* itoa(long long value, char * str, int base);
 int printchar(char c);
 char* gets();
 void stdinput(char);
